# Faster implementation of rabbit_binding:recover/0