For a large number of bindings, this implementation reduces `rabbit_binding:recover/0`
execution from minutes to seconds. rabbit_binding:recover/2 can still take minutes but it can be improved separately.

Before:
```
[debug] <0.9.0> Time to start RabbitMQ: 136651247 µs
```
After:
```
[debug] <0.222.0> rabbit_binding:recover/0 completed in 1.451836s
[debug] <0.436.0> rabbit_binding:recover/2 for vhost / completed in 51.545797s
[debug] <0.9.0> Time to start RabbitMQ: 83644782 µs
```
That's 136s to 83s for a test with 16k queues and 275k bindings.