@michaelklishin 2021-06-29
Backported to `v3.8.x`.