@michaelklishin 2021-06-29
Backported to `v3.9.x`